+++ moveit_simple/CMakeLists.txt
@@ -58,10 +58,16 @@ include_directories(
   ${EIGEN3_INCLUDE_DIRS}
 )
 
-add_library(${PROJECT_NAME} src/moveit_simple.cpp)
+add_library(${PROJECT_NAME} 
+  src/online_robot.cpp
+  src/point_types.cpp
+  src/robot.cpp
+)
+
 add_dependencies(${PROJECT_NAME} ${catkin_EXPORTED_TARGETS})
 add_dependencies(${PROJECT_NAME} ${PROJECT_NAME}_gencfg)
 add_dependencies(${PROJECT_NAME} ${PROJECT_NAME}_genparam)
+
 target_link_libraries(${PROJECT_NAME} ${catkin_LIBRARIES})
 
 if(CATKIN_ENABLE_TESTING)
