+++ moveit_simple/src/robot.cpp
@@ -1,7 +1,8 @@
 /*
  * Software License Agreement (Apache License)
  *
- * Copyright (c) 2016, Shaun Edwards
+ * Copyright (c) 2016 Shaun Edwards
+ * Copyright (c) 2018 Plus One Robotics
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,10 +17,18 @@
  * limitations under the License.
  */
 
+#include <map>
+#include <string>
+#include <vector>
+
 #include <eigen_conversions/eigen_msg.h>
-#include <moveit_simple/moveit_simple.h>
-#include <prettyprint/prettyprint.hpp>
-#include <ros/console.h>
+#include <ros/ros.h>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
+
+
+#include <moveit_simple/exceptions.h>
+#include <moveit_simple/point_types.h>
+#include <moveit_simple/robot.h>
 
 namespace moveit_simple
 {
@@ -88,54 +97,6 @@ Robot::Robot(const ros::NodeHandle &nh, const std::string &robot_description,
 }
 
 Robot::~Robot() { }
-
-OnlineRobot::OnlineRobot(const ros::NodeHandle &nh, const std::string &robot_description,
-  const std::string &group_name) : Robot(nh, robot_description, group_name), 
-  action_("joint_trajectory_action", true)
-{
-  current_robot_state_.reset(new moveit::core::RobotState(robot_model_ptr_));
-  current_robot_state_->setToDefaultValues();
-
-  online_visual_tools_.reset(new moveit_visual_tools::MoveItVisualTools(robot_model_ptr_->getRootLinkName(),
-    nh_.getNamespace() + "/rviz_visual_tools", robot_model_ptr_));
-  online_visual_tools_->loadRobotStatePub(nh_.getNamespace() + "/display_robot_state");
-
-  ROS_INFO_STREAM("Waiting for action servers");
-  action_.waitForServer(ros::Duration(30.0));
-  ROS_INFO_STREAM("Done waiting for action servers");
-
-  ROS_INFO_STREAM("Loading ROS pubs/subs");
-  j_state_sub_ = nh_.subscribe("joint_states", 1, &OnlineRobot::updateCurrentState, this);
-
-  // TODO: How to handle action server and other failures in the constructor
-  // Perhaps move any items that can fail our of the constructor into an init
-  // function with a proper return
-  if (!action_.isServerConnected())
-  {
-    ROS_ERROR_STREAM("Failed to connect to joint trajectory action server: ");
-  }
-
-  return;
-}
-
-OnlineRobot::OnlineRobot(const ros::NodeHandle &nh, const std::string &robot_description,
-  const std::string &group_name, const std::string &ik_base_frame, const std::string &ik_tip_frame)
-  : OnlineRobot(nh, robot_description, group_name)
-{
-  try
-  {
-    ik_base_frame_ = ik_base_frame;
-    ik_tip_frame_ = ik_tip_frame;
-    this->computeIKSolverTransforms();
-  }
-  catch (tf2::TransformException &ex)
-  {
-    ROS_ERROR_STREAM("Failed to compute transforms between the base/tip frames defined"
-      << " in the SRDF and the base/tip frames defined for the IK solver");
-    throw IKSolverTransformException("Failed to compute transforms between the base/tip"
-      " frame defined in the SRDF and the base/tip frames defined for the IK solver");
-  }
-}
 
 void Robot::addTrajPoint(const std::string &traj_name, const Eigen::Affine3d pose,
   const std::string &frame, double time, const InterpolationType &type,
@@ -686,75 +647,6 @@ std::vector<moveit_simple::JointTrajectoryPoint> Robot::plan(const std::string t
   {
     ROS_ERROR_STREAM("Trajectory [" << traj_name << "] not found");
     throw std::invalid_argument("No trajectory found named " + traj_name);
-  }
-}
-
-void OnlineRobot::execute(const std::string traj_name, bool collision_check)
-{
-  std::lock_guard<std::recursive_mutex> guard(m_);
-
-  try
-  {
-    std::vector<moveit_simple::JointTrajectoryPoint> goal;
-
-    goal = plan(traj_name, collision_check);
-    execute(goal);
-  }
-  catch (IKFailException &ik)
-  {
-    ROS_ERROR_STREAM("IK Failed for trajectory: [" << traj_name << "]");
-    throw ik;
-  }
-  catch (CollisionDetected &cd)
-  {
-    ROS_ERROR_STREAM("Collision detected in trajectory");
-    throw cd;
-  }
-  catch (std::invalid_argument &ia)
-  {
-    ROS_ERROR_STREAM("Invalid trajectory name: [" << traj_name << "]");
-    throw ia;
-  }
-  catch (ExecutionFailureException &ef)
-  {
-    ROS_ERROR_STREAM("Trajectory [" << traj_name << "] failed to execute");
-    throw ef;
-  }
-}
-
-void OnlineRobot::execute(std::vector<moveit_simple::JointTrajectoryPoint> &joint_trajectory_points,
-  bool collision_check)
-{
-  std::lock_guard<std::recursive_mutex> guard(m_);
-
-  const double TIMEOUT_SCALE = 1.25;  // scales time to wait for action timeout.
-
-  control_msgs::FollowJointTrajectoryGoal goal = toFollowJointTrajectoryGoal(joint_trajectory_points);
-
-  int collision_points = trajCollisionCheck(goal, collision_check);
-
-  if (collision_points == 0)
-  {
-    ros::Duration traj_time =
-      goal.trajectory.points[goal.trajectory.points.size() - 1].time_from_start;
-
-    ros::Duration timeout(TIMEOUT_SCALE * traj_time.toSec());
-
-    if (action_.sendGoalAndWait(goal, timeout) == actionlib::SimpleClientGoalState::SUCCEEDED)
-    {
-      ROS_INFO_STREAM("Successfully executed Joint Trajectory ");
-    }
-    else
-    {
-      ROS_ERROR_STREAM("Joint Trajectory failed to execute.. check input Plan");
-      throw ExecutionFailureException("Execution failed for Joint Trajectory");
-    }
-  }
-  else
-  {
-    ROS_ERROR_STREAM("Collision detected at " << collision_points 
-      << " points for Joint Trajectory");
-    throw CollisionDetected("Collision detected while interpolating ");
   }
 }
 
@@ -1355,68 +1247,4 @@ std::vector<double> Robot::getJointState(void) const
 
   return current_joint_positions;
 }
-
-std::vector<double> OnlineRobot::getJointState(void) const
-{
-  std::lock_guard<std::recursive_mutex> guard(m_);
-
-  ros::spinOnce();
-  std::vector<double> current_joint_positions;
-
-  current_robot_state_->update();
-  current_robot_state_->copyJointGroupPositions(joint_group_->getName(), current_joint_positions);
-
-  return current_joint_positions;
-}
-
-void OnlineRobot::updateCurrentState(const sensor_msgs::JointStateConstPtr &msg)
-{
-  current_robot_state_->setVariablePositions(msg->name, msg->position);
-}
-
-std::unique_ptr<JointTrajectoryPoint> JointTrajectoryPoint::toJointTrajPoint(
-  const Robot &robot, double timeout, const std::vector<double> &seed) const
-{
-  ROS_DEBUG_STREAM("JointTrajectoryPoint: passing through joint trajectory point");
-  return std::unique_ptr<JointTrajectoryPoint>(new JointTrajectoryPoint(*this));
-}
-
-std::unique_ptr<CartTrajectoryPoint> JointTrajectoryPoint::toCartTrajPoint(const Robot &robot) const
-{
-  Eigen::Affine3d pose;
-
-  ROS_DEBUG_STREAM("JointTrajectoryPoint: Calculating FK for Cartesian trajectory point");
-  if (robot.getPose(joint_point_, pose))
-  {
-    return std::unique_ptr<CartTrajectoryPoint>(new CartTrajectoryPoint(pose, time(), name()));
-  }
-  else
-  {
-    ROS_WARN_STREAM("Failed to find FK for point: " << name_);
-    return std::unique_ptr<CartTrajectoryPoint>(nullptr);
-  }
-}
-
-std::unique_ptr<JointTrajectoryPoint> CartTrajectoryPoint::toJointTrajPoint(
-  const Robot &robot, double timeout, const std::vector<double> &seed) const
-{
-  std::vector<double> joints;
-
-  ROS_DEBUG_STREAM("CartTrajectoryPoint: Calculating IK for joint trajectory point");
-  if (robot.getJointSolution(pose_, timeout, seed, joints))
-  {
-    return std::unique_ptr<JointTrajectoryPoint>(new JointTrajectoryPoint(joints, time(), name()));
-  }
-  else
-  {
-    ROS_WARN_STREAM("Failed to find joint solution for point: " << name_);
-    return std::unique_ptr<JointTrajectoryPoint>(nullptr);
-  }
-}
-
-std::unique_ptr<CartTrajectoryPoint> CartTrajectoryPoint::toCartTrajPoint(const Robot &robot) const
-{
-  ROS_DEBUG_STREAM("CartTrajectoryPoint: passing through cartesian trajectory point");
-  return std::unique_ptr<CartTrajectoryPoint>(new CartTrajectoryPoint(*this));
-}
-}
+} // namespace moveit_simple
