+++ moveit_simple/test/motoman_mh5_utest.cpp
@@ -1,7 +1,8 @@
 /*
  * Software License Agreement (Apache License)
  *
- * Copyright (c) 2014, Southwest Research Institute
+ * Copyright (c) 2016 Shaun Edwards
+ * Copyright (c) 2018 Plus One Robotics
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
